+++ ff/src/biginteger/mod.rs
@@ -4,6 +4,7 @@ use crate::{
     fields::{BitIteratorBE, BitIteratorLE},
     UniformRand,
 };
+use ark_ff_macros::unroll_for_loops;
 use ark_serialize::{CanonicalDeserialize, CanonicalSerialize, SerializationError};
 use ark_std::{
     convert::TryFrom,
@@ -74,7 +75,7 @@ macro_rules! BigInt {
 
 #[doc(hidden)]
 macro_rules! const_modulo {
-    ($a:ident, $divisor:ident) => {{
+    ($a:expr, $divisor:expr) => {{
         // Stupid slow base-2 long division taken from
         // https://en.wikipedia.org/wiki/Division_algorithm
         assert!(!$divisor.const_is_zero());
@@ -85,7 +86,7 @@ macro_rules! const_modulo {
             remainder = remainder.const_mul2();
             remainder.0[0] |= $a.get_bit(i as usize) as u64;
             if remainder.const_geq($divisor) {
-                let (r, borrow) = remainder.const_sub_noborrow($divisor);
+                let (r, borrow) = remainder.const_sub_with_borrow($divisor);
                 remainder = r;
                 assert!(!borrow);
             }
@@ -94,6 +95,7 @@ macro_rules! const_modulo {
         remainder
     }};
 }
+
 impl<const N: usize> BigInt<N> {
     #[doc(hidden)]
     pub const fn const_is_even(&self) -> bool {
@@ -134,7 +136,7 @@ impl<const N: usize> BigInt<N> {
         true
     }
 
-    /// Compute the largest integer `s` such that `self = 2**s * t` for odd `t`.
+    /// Compute the largest integer `s` such that `self = 2**s * t + 1` for odd `t`.
     #[doc(hidden)]
     pub const fn two_adic_valuation(mut self) -> u32 {
         let mut two_adicity = 0;
@@ -149,7 +151,7 @@ impl<const N: usize> BigInt<N> {
         two_adicity
     }
 
-    /// Compute the smallest odd integer `t` such that `self = 2**s * t` for some
+    /// Compute the smallest odd integer `t` such that `self = 2**s * t + 1` for some
     /// integer `s = self.two_adic_valuation()`.
     #[doc(hidden)]
     pub const fn two_adic_coefficient(mut self) -> Self {
@@ -182,8 +184,8 @@ impl<const N: usize> BigInt<N> {
     }
 
     #[inline]
-    #[ark_ff_asm::unroll_for_loops]
-    pub(crate) const fn const_sub_noborrow(mut self, other: &Self) -> (Self, bool) {
+    #[unroll_for_loops(12)]
+    pub(crate) const fn const_sub_with_borrow(mut self, other: &Self) -> (Self, bool) {
         let mut borrow = 0;
 
         const_for!((i in 0..N) {
@@ -205,7 +207,7 @@ impl<const N: usize> BigInt<N> {
         self
     }
 
-    #[ark_ff_asm::unroll_for_loops]
+    #[unroll_for_loops(12)]
     pub(crate) const fn const_is_zero(&self) -> bool {
         let mut is_zero = true;
         crate::const_for!((i in 0..N) {
@@ -234,8 +236,8 @@ impl<const N: usize> BigInteger for BigInt<N> {
     const NUM_LIMBS: usize = N;
 
     #[inline]
-    #[ark_ff_asm::unroll_for_loops]
-    fn add_nocarry(&mut self, other: &Self) -> bool {
+    #[unroll_for_loops(12)]
+    fn add_with_carry(&mut self, other: &Self) -> bool {
         let mut carry = 0;
 
         for i in 0..N {
@@ -256,8 +258,8 @@ impl<const N: usize> BigInteger for BigInt<N> {
     }
 
     #[inline]
-    #[ark_ff_asm::unroll_for_loops]
-    fn sub_noborrow(&mut self, other: &Self) -> bool {
+    #[unroll_for_loops(12)]
+    fn sub_with_borrow(&mut self, other: &Self) -> bool {
         let mut borrow = 0;
 
         for i in 0..N {
@@ -278,7 +280,7 @@ impl<const N: usize> BigInteger for BigInt<N> {
     }
 
     #[inline]
-    #[ark_ff_asm::unroll_for_loops]
+    #[unroll_for_loops(12)]
     #[allow(unused)]
     fn mul2(&mut self) {
         #[cfg(all(target_arch = "x86_64", feature = "asm"))]
@@ -308,7 +310,7 @@ impl<const N: usize> BigInteger for BigInt<N> {
     }
 
     #[inline]
-    #[ark_ff_asm::unroll_for_loops]
+    #[unroll_for_loops(12)]
     fn muln(&mut self, mut n: u32) {
         if n >= (64 * N) as u32 {
             *self = Self::from(0u64);
@@ -337,7 +339,7 @@ impl<const N: usize> BigInteger for BigInt<N> {
     }
 
     #[inline]
-    #[ark_ff_asm::unroll_for_loops]
+    #[unroll_for_loops(12)]
     #[allow(unused)]
     fn div2(&mut self) {
         let mut t = 0;
@@ -351,7 +353,7 @@ impl<const N: usize> BigInteger for BigInt<N> {
     }
 
     #[inline]
-    #[ark_ff_asm::unroll_for_loops]
+    #[unroll_for_loops(12)]
     fn divn(&mut self, mut n: u32) {
         if n >= (64 * N) as u32 {
             *self = Self::from(0u64);
@@ -522,7 +524,7 @@ impl<const N: usize> Display for BigInt<N> {
 
 impl<const N: usize> Ord for BigInt<N> {
     #[inline]
-    #[ark_ff_asm::unroll_for_loops]
+    #[unroll_for_loops(12)]
     fn cmp(&self, other: &Self) -> ::core::cmp::Ordering {
         use core::cmp::Ordering;
         for i in 0..N {
@@ -671,6 +673,7 @@ mod tests;
 
 /// This defines a `BigInteger`, a smart wrapper around a
 /// sequence of `u64` limbs, least-significant limb first.
+// TODO: get rid of this trait once we can use associated constants in const generics.
 pub trait BigInteger:
     ToBytes
     + FromBytes
@@ -701,42 +704,48 @@ pub trait BigInteger:
     /// Number of 64-bit limbs representing `Self`.
     const NUM_LIMBS: usize;
 
-    /// Add another representation to this one, returning the carry bit.
+    /// Add another [`BigInteger`] to `self`. This method stores the result in `self`,
+    /// and returns a carry bit.
+    ///
     /// # Example
     ///
     /// ```
     /// use ark_ff::{biginteger::BigInteger64 as B, BigInteger as _};
     ///
     /// // Basic
-    /// let (mut one, mut two_add) = (B::from(1u64), B::from(2u64));
-    /// two_add.add_nocarry(&one);
-    /// assert_eq!(two_add, B::from(3u64));
+    /// let (mut one, mut x) = (B::from(1u64), B::from(2u64));
+    /// let carry = x.add_with_carry(&one);
+    /// assert_eq!(x, B::from(3u64));
+    /// assert_eq!(carry, false);
     ///
     /// // Edge-Case
-    /// let mut max_one = B::from(u64::MAX);
-    /// let carry = max_one.add_nocarry(&one);
-    /// assert_eq!(max_one, B::from(0u64));
+    /// let mut x = B::from(u64::MAX);
+    /// let carry = x.add_with_carry(&one);
+    /// assert_eq!(x, B::from(0u64));
     /// assert_eq!(carry, true)
     /// ```
-    fn add_nocarry(&mut self, other: &Self) -> bool;
+    fn add_with_carry(&mut self, other: &Self) -> bool;
 
-    /// Subtract another representation from this one, returning the borrow bit.
+    /// Subtract another [`BigInteger`] from this one. This method stores the result in
+    /// `self`, and returns a borrow.
+    ///
     /// # Example
     ///
     /// ```
     /// use ark_ff::{biginteger::BigInteger64 as B, BigInteger as _};
     ///
     /// // Basic
-    /// let (mut one, mut two, mut three_sub) = (B::from(1u64), B::from(2u64), B::from(3u64));
-    /// three_sub.sub_noborrow(&two);
-    /// assert_eq!(three_sub, one);
+    /// let (mut one_sub, two, mut three_sub) = (B::from(1u64), B::from(2u64), B::from(3u64));
+    /// let borrow = three_sub.sub_with_borrow(&two);
+    /// assert_eq!(three_sub, one_sub);
+    /// assert_eq!(borrow, false);
     ///
     /// // Edge-Case
-    /// let borrow = one.sub_noborrow(&two);
+    /// let borrow = one_sub.sub_with_borrow(&two);
+    /// assert_eq!(one_sub, B::from(u64::MAX));
     /// assert_eq!(borrow, true);
-    /// assert_eq!(one, B::from(u64::MAX));
     /// ```
-    fn sub_noborrow(&mut self, other: &Self) -> bool;
+    fn sub_with_borrow(&mut self, other: &Self) -> bool;
 
     /// Performs a leftwise bitshift of this number, effectively multiplying
     /// it by 2. Overflow is ignored.
@@ -763,7 +772,8 @@ pub trait BigInteger:
     /// ```
     fn mul2(&mut self);
 
-    /// Performs a leftwise bitshift of this number by some amount.
+    /// Performs a leftwise bitshift of this number by n bits, effectively multiplying
+    /// it by 2^n. Overflow is ignored.
     /// # Example
     ///
     /// ```
@@ -996,9 +1006,9 @@ pub trait BigInteger:
                 if e.is_odd() {
                     z = signed_mod_reduction(e.as_ref()[0], 1 << w);
                     if z >= 0 {
-                        e.sub_noborrow(&Self::from(z as u64));
+                        e.sub_with_borrow(&Self::from(z as u64));
                     } else {
-                        e.add_nocarry(&Self::from((-z) as u64));
+                        e.add_with_carry(&Self::from((-z) as u64));
                     }
                 } else {
                     z = 0;